requirements.txt
cffi==1.6.0
cryptography==1.4
idna==2.1
ndg-httpsclient==0.4.1
nose==1.3.7
Pillow==3.2.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
requests==2.10.0
six==1.10.0
